SeriesAcceleration

TODO: desciption here

Index

SeriesAcceleration.acc_csum

SeriesAcceleration.build_M_matrix

SeriesAcceleration.build_weights_bender

SeriesAcceleration.build_weights_rohringer

SeriesAcceleration.rateOfConv

SeriesAcceleration.acc_csum — Method

acc_sum(arr::AbstractArray{T1,1}, type::T2) where {T1 <: Number, T2 <: SumHelper}

arr specifies a 1D array of partial sums using a method specified by type

SeriesAcceleration.build_M_matrix — Method

build_M_matrix(dom::AbstractArray{Int,1}, exponents::AbstractArray{Int,1})

Helper function that build the matrix $M$ used to fit data obtained for indices dom to a $\sum_{n\in exponents} c_n/i^n$ tail.

dom specifies the number of terms for the partial sum, that are to be fitted. exponents specifies the exponents of $\sum_{x \in dom} \sum_{p in exponents} 1/i^p$ that the partial sums should be fitted to.

The coefficients $c_n$ are obtained by solving $M c = b$. $b$ can be constructed from the data using build_rhs.

SeriesAcceleration.build_weights_bender — Method

buildweightsbender(nstart::Int, dom::AbstractArray{Int,1}, exponents::AbstractArray{Int,1})

Build weight matrix in closed form. See C. Bender, A. Orszag 99, p. 375. Fit coefficients can be obtained by multiplying w with data: $a_k = W_{kj} g_j$

SeriesAcceleration.build_weights_rohringer — Method

buildweightsrohringer(dom::AbstractArray{Int,1}, exponents::AbstractArray{Int,1})

Build weight matrix i.e. $W = M^{-1} R$ with M from build_M_matrix and $R_{kj} = \frac{1}{j^k}$. Fit coefficients can be obtained by multiplying w with data: $a_k = W_{kj} g_j$

SeriesAcceleration.rateOfConv — Method

rateOfConv(arr:AbstractArray{T, 1})

Estimates rate of convergence $\alpha_n = \frac{\log |(x_{n+1} - x_n)/(x_n - x_{n-1})|}{\log |(x_n - x_{n-1})/(x_{n-1}-x_{n-2})|}$ from array. trace can be set to true, to obtain the roc for all partial sums.